docs(oops): complete 4-gate verification across all 21 chapters
- Synchronize Readme.txt and create Readme.md for Chapters 28, F, 5, 10, 11, 12, 13
- Add missing code examples for trailing return types, stream tie/sync, binary file I/O, constinit, if consteval
- Fix pedagogical compilation errors with diagnostic comments across code examples
- Remove redundant OOPs/AGENTS.md in favor of canonical global steering rule
- Update chapter progress tracker in learncpp_oops_folder_work.md

+++ OOPs/10_TypeConversion_TypeAliases_TypeDeduction/10_3_Narrowing_conversions_list_initialization_constexpr_initializers/3_int_to_float.cpp
@@ -8,9 +8,10 @@ int main()
     float y{x};
 */
 
-    // narrowing conversion as x is not constexpr and x int value cant be fit in y float.
+    // Narrowing conversion: constexpr int value 16777217 (2^24 + 1) cannot fit exactly into 24-bit float mantissa
     constexpr int x = 16777217;
-    float y{x};
+    // float y{x}; // COMPILE ERROR: narrowing conversion in list-initialization
+    float y{ static_cast<float>(x) }; // OK: explicit static_cast acknowledges loss of precision
 
     std::cout << "y : " << std::fixed << y << std::endl;
     return 0;
